% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\alias{fill_natal}
\alias{fill_regional}
\alias{fill_natal_dens_depend}
\title{Filling Natal Tributary and Regional Habitat}
\source{
IP-117068
}
\usage{
fill_natal(
  juveniles,
  inchannel_habitat,
  floodplain_habitat,
  territory_size = springRunDSM::params$territory_size,
  up_to_size_class = 2,
  yearlings = FALSE
)

fill_regional(
  juveniles,
  habitat,
  floodplain_habitat = NULL,
  territory_size = springRunDSM::params$territory_size,
  up_to_size_class = 3
)

fill_natal_dens_depend(
  juveniles,
  inchannel_habitat,
  floodplain_habitat,
  up_to_size_class = 2,
  ..floodplain_capacity = 5,
  ..habitat_capacity = 4,
  territory_size = territory_size,
  ...
)
}
\arguments{
\item{juveniles}{An n by 4 matrix of juvenile fish size s, m, l, vl}

\item{inchannel_habitat}{The available tributary habitat in square meters}

\item{floodplain_habitat}{The available floodplain habitat in square meters (NULL for bypasses and delta)}

\item{territory_size}{A length 4 array of juvenile fish territory requirements s, m, l, vl}

\item{up_to_size_class}{Which size class and under is moved to floodplain and
inchannel habitat (small = 1, medium = 2, large = 3, very large = 4)}
}
\description{
Allocates juvenile fish onto the floodplain and in-channel habitat. Once
available habitat is full fish are assigned to out migrate.
The \code{fill_natal} function is used to route fish within their natal tributary
and the \code{fill_regional} function is used to route fish on the mainstem, bypasses, and deltas.
}
\details{
See \code{\link{params}} for details on parameter sources
}
