% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_month_dynamic.R
\name{juvenile_month_dynamic}
\alias{juvenile_month_dynamic}
\title{Juvenile Month Dynamics}
\usage{
juvenile_month_dynamic(
  fish,
  year = year,
  month = month,
  mode,
  rearing_survival = rearing_survival,
  migratory_survival = migratory_survival,
  habitat = habitat,
  ..params = ..params,
  avg_ocean_transition_month = avg_ocean_transition_month,
  stochastic = stochastic,
  ic_growth,
  fp_growth,
  delta_growth,
  filling_fn = fill_natal,
  filling_args = NULL,
  filling_regional_fn = fill_regional,
  filling_regional_args = NULL,
  movement_fn = NULL,
  movement_args = NULL,
  movement_months = NULL
)
}
\arguments{
\item{fish}{list object containing tracking matrices for juveniles, yearlings, juveniles at chipps, and adults in ocean}

\item{year}{year}

\item{month}{month}

\item{mode}{mode}

\item{rearing_survival}{rearing survival for year and month}

\item{migratory_survival}{migratory survival for year and month}

\item{habitat}{habitat for year and month}

\item{..params}{full params list from model parms}

\item{hypothesis}{movement hypothesis to test}
}
\description{
Function to run through monthly juvenile routing and rearing logic.
}
\examples{
fish = list(juveniles = juveniles,
            yearlings = yearlings,
            north_delta_fish = north_delta_fish,
            south_delta_fish = south_delta_fish,
            juveniles_at_chipps = juveniles_at_chipps,
            adults_in_ocean = adults_in_ocean)
}
