% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spring_hatchery_release}
\alias{spring_hatchery_release}
\title{Spring Hatchery Release}
\format{
A matrix with dimensions 31 x 4 x 20 (watershed x size class x simulation year)
}
\source{
Hatchery return data comes from the production targets defined in the
\href{https://swfsc-publications.fisheries.noaa.gov/publications/CR/2012/2012California.pdf}{California HSRG (Page 82)}
}
\usage{
spring_hatchery_release
}
\description{
Hatchery release by size class and hatchery
}
\keyword{datasets}
