% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{spring_run_model}
\alias{spring_run_model}
\title{Spring Run Chinook Model}
\source{
IP-117068
}
\usage{
spring_run_model(scenario = NULL, seeds = NULL)
}
\arguments{
\item{scenario}{Model inputs, can be modified to test management actions}

\item{seeds}{The default value is NULL runs the model in seeding mode,
returning a 31 by 25 matrix with the first four years of seeded adults. This
returned value can be fed into the model again as the value for the seeds argument}
}
\description{
Spring Run Chinook life cycle model used for CVPIA's Structured
Decision Making Process
}
