% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlings.R
\name{yearling_growth}
\alias{yearling_growth}
\title{Spring-Run Yearlings}
\usage{
yearling_growth(
  year,
  yearlings,
  territory_size = c(0.05423379, 0.14539419, 0.48471474, 0.48471474)
)
}
\description{
Spring-Run Yearlings
}
