% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlings.R
\name{yearling_migration}
\alias{yearling_migration}
\title{Yearling Migration}
\usage{
yearling_migration(
  year,
  yearlings,
  migrants,
  north_delta_fish,
  south_delta_fish,
  juveniles_at_chipps,
  growth_rates,
  annual_migrants,
  avg_ocean_transition_month,
  month = 11
)
}
\description{
Yearling Migration
}
